refactor(api-bridge): split actor action handler into per-action files
